requirements.txt
apache-airflow==2.9.3
pandas
mysql-connector-python
